Sas tokens needs to be cached in a singleton. If scoped, caching does not work.
