+++ src/Altinn.Correspondence.Integrations/DependencyInjection.cs
@@ -31,6 +31,7 @@ public static class DependencyInjection
         services.AddScoped<IResourceManager, AzureResourceManagerService>();
         services.AddScoped<IStorageConnectionStringRepository, AzureResourceManagerService>();
         services.AddScoped<IResourceRegistryService, ResourceRegistryService>();
+        services.AddSingleton<SasTokenCacheService, SasTokenCacheService>();
         if (string.IsNullOrWhiteSpace(maskinportenSettings.ClientId))
         {
             services.AddScoped<IEventBus, ConsoleLogEventBus>();
